requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pymupdf
langchain-community
python-dotenv
-e .